Clean up old files and queries
