+++ src/components/daoCard/DaoCard.js
@@ -1,26 +1,16 @@
 import React, { useContext } from 'react';
-import { useQuery } from '@apollo/react-hooks';
 
-import { GET_MEMBERDATA } from '../../util/queries';
 import { Web3Context } from '../../contexts/ContractContexts';
 
 import './DaoCard.scss';
 
-const DaoCard = props => {
-  const { dao } = props;
+const DaoCard = ({ dao }) => {
   const [web3Service] = useContext(Web3Context);
-
-  const { loading, error, data } = useQuery(GET_MEMBERDATA, {
-    variables: { contractAddr: dao.moloch },
-  });
 
   const bankValue = value => {
     const amt = web3Service.fromWei(value);
     return parseFloat(amt).toFixed(2);
   };
-
-  if (loading) return <p>Loading...</p>;
-  if (error) return <p>Error - use mainnet :(</p>;
 
   return (
     <>
@@ -28,11 +18,17 @@ const DaoCard = props => {
         <div className="DaoCard">
           <h4 className="DaoName">{dao.apiData.name || dao.title}</h4>
           <p>{dao.apiData.description}</p>
+          <p>Moloch Version {dao.version}</p>
+
           <div className="Row">
             <div className="Column">
+              <>
+                <p className="Label">Shares</p>
+                <p className="Data">{dao.totalShares}</p>
+              </>
               {dao.version !== '2' ? (
                 <>
-                  <p className="Label">Bank</p>
+                  <p className="Label">Bank Value</p>
                   <p className="Data">
                     {bankValue(dao.tokenInfo.guildBankValue)}{' '}
                     {dao.tokenInfo.symbol}
@@ -40,22 +36,17 @@ const DaoCard = props => {
                 </>
               ) : (
                 <>
-                  <p className="Label">Shares</p>
-                  <p className="Data">{dao.metadata.totalShares}</p>
+                  <p className="Label">Bank Tokens</p>
+                  <p className="Data">{dao.approvedTokens.length}</p>
                 </>
               )}
             </div>
             <div className="Column">
               <p className="Label">Members</p>
-              {dao.version !== '2' ? (
-                <p className="Data">
-                  {dao.apiData.legacyData
-                    ? dao.apiData.legacyData.members.length
-                    : data.members.length}
-                </p>
-              ) : (
-                <p className="Data">{dao.metadata.members.length}</p>
-              )}
+              <p className="Data">{dao.members.length}</p>
+
+              <p className="Label">Proposals</p>
+              <p className="Data">{dao.proposals.length}</p>
             </div>
           </div>
         </div>
