+++ src/components/daoList/DaoList.js
@@ -1,40 +1,16 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 
 import DaoCard from '../daoCard/DaoCard';
 
 import './DaoList.scss';
 
-const DaoList = props => {
-  const { daos, version, v2Moloches } = props;
-  const [mergedData, setMergedData] = useState([]);
-
-  useEffect(() => {
-    if (daos) {
-      setMergedData(
-        version === '1'
-          ? daos
-          : daos.map((dao, i) => {
-              dao.metadata = v2Moloches.find(v2 => {
-                return v2.id === dao.id;
-              });
-              return dao;
-            }),
-      );
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [daos]);
-
-  const daoList = mergedData.map(dao => {
+const DaoList = ({ daos }) => {
+  const daoList = daos.map(dao => {
     return (
-      <div className="DaoList__Item" key={dao.moloch}>
+      <div className="DaoList__Item" key={dao.id}>
         <Link
-          to={
-            dao.version === '2'
-              ? `/dao/v2/${dao.moloch}`
-              : `/dao/v1/${dao.moloch}`
-          }
+          to={dao.version === '2' ? `/dao/v2/${dao.id}` : `/dao/v1/${dao.id}`}
         >
           <DaoCard dao={dao} />
         </Link>
@@ -44,7 +20,7 @@ const DaoList = props => {
 
   return (
     <div className="Contain">
-      {mergedData ? <div className="DaoList">{daoList}</div> : null}
+      {daos ? <div className="DaoList">{daoList}</div> : null}
     </div>
   );
 };
