+++ src/util/queries.js
@@ -1,115 +1,5 @@
 import { gql } from 'apollo-boost';
 
-export const GET_MEMBERDATA = gql`
-  query members($contractAddr: String!) {
-    members(where: { molochAddress: $contractAddr }) {
-      id
-      delegateKey
-      molochAddress
-      shares
-      memberId @client
-    }
-  }
-`;
-
-export const GET_MEMBERDATA_LEGACY = {
-  query: 'query members { members { id delegateKey shares }}',
-};
-
-export const GET_MOLOCHES_POST = {
-  query: `query {
-    factories(orderBy: index) {
-      id
-    }
-  }`,
-};
-
-export const GET_MOLOCHES_POST_V2 = {
-  query: `query {
-    daos(orderBy: index) {
-      id
-    }
-  }`,
-};
-
-export const GET_MOLOCHES = gql`
-  query factories($skip: Int) {
-    factories(orderBy: index, first: 100, skip: $skip) {
-      apiData @client
-      id
-      title
-      moloch
-      summoner
-      tokenInfo @client
-    }
-  }
-`;
-
-export const GET_MOLOCHES_V2 = gql`
-  query daos($skip: Int) {
-    daos(orderBy: index, first: 100, skip: $skip) {
-      apiData @client
-      id
-      index
-      moloch
-      summoner
-      version
-      title
-    }
-    moloches(orderBy: summoningTime, first: 100, skip: $skip) {
-      id
-      totalShares
-      summoningTime
-      members(where: { exists: true }) {
-        id
-      }
-    }
-  }
-`;
-
-export const GET_MOLOCH = gql`
-  query factories($contractAddr: String!) {
-    factories(where: { id: $contractAddr }) {
-      apiData @client
-      id
-      title
-      moloch
-      summoner
-      newContract
-      tokenInfo @client
-      totalShares @client
-    }
-  }
-`;
-
-export const GET_MOLOCH_V2 = gql`
-  query daos($contractAddr: String!) {
-    daos(where: { id: $contractAddr }) {
-      apiData @client
-      id
-      title
-      moloch
-      summoner
-      version
-    }
-    moloches(where: { id: $contractAddr }) {
-      id
-      totalShares
-      summoningTime
-      members(where: { exists: true }) {
-        id
-        memberAddress
-        delegateKey
-        shares
-        loot
-      }
-      depositToken {
-        tokenAddress
-      }
-    }
-  }
-`;
-
 export const GET_MOLOCHES_STATS = gql`
   query {
     factories(orderBy: index) {
@@ -126,24 +16,7 @@ export const GET_MOLOCHES_STATS = gql`
   }
 `;
 
-export const GET_PROPOSALS = gql`
-  query proposals($contractAddr: String!) {
-    proposals(where: { molochAddress: $contractAddr }) {
-      id
-      timestamp
-      molochAddress
-      votes {
-        id
-      }
-    }
-  }
-`;
-
-export const GET_PROPOSALS_LEGACY = {
-  query: 'query proposals { proposals { id, timestamp, votes { id } }}',
-};
-
-export const GET_SUPER_MOLOCHES = gql`
+export const GET_MOLOCHES = gql`
   query moloches($skip: Int) {
     moloches(orderBy: summoningTime, first: 100, skip: $skip) {
       apiData @client
@@ -166,7 +39,7 @@ export const GET_SUPER_MOLOCHES = gql`
   }
 `;
 
-export const GET_SUPER_MOLOCH = gql`
+export const GET_MOLOCH = gql`
   query moloch($contractAddr: String!) {
     moloch(id: $contractAddr) {
       apiData @client
@@ -224,7 +97,7 @@ export const GET_MEMBER_GAS = gql`
 
 export const GET_MEMBER_MOLOCHES = gql`
   query members($memberAddress: String!) {
-    members(where: { memberAddress: $memberAddress }) {
+    members(where: { memberAddress: $memberAddress, exists: true }) {
       id
       moloch {
         apiData @client
