+++ src/views/stats/Stats.js
@@ -1,58 +1,69 @@
 import React from 'react';
-import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
-import { useQuery } from '@apollo/react-hooks';
-import _ from 'lodash';
+// import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
+// import { useQuery } from '@apollo/react-hooks';
+// import _ from 'lodash';
 
-import { GET_MOLOCHES_STATS } from '../../util/queries';
+// import { GET_MOLOCHES_STATS } from '../../util/queries';
 
 import './Stats.scss';
-import GuildBanks from '../../components/stats/GuildBank';
-import ProposalStats from '../../components/stats/ProposalStats';
-import Activity from '../../components/stats/Activity';
+// import GuildBanks from '../../components/stats/GuildBank';
+// import ProposalStats from '../../components/stats/ProposalStats';
+// import Activity from '../../components/stats/Activity';
 
 const Stats = props => {
-  const { loading, error, data } = useQuery(GET_MOLOCHES_STATS);
+  // const { loading, error, data } = useQuery(GET_MOLOCHES_STATS);
 
-  const totalDaos = () => {
-    return data.factories.length;
-  };
+  // const totalDaos = () => {
+  //   return data.factories.length;
+  // };
 
-  const totalShares = () => {
-    return data.factories.reduce((sum, dao) => {
-      return sum + +dao.totalShares;
-    }, 0);
-  };
+  // const totalShares = () => {
+  //   return data.factories.reduce((sum, dao) => {
+  //     return sum + +dao.totalShares;
+  //   }, 0);
+  // };
 
-  const totalDoaMemberships = () => {
-    const counts = data.factories.reduce(
-      (sum, dao) => {
-        +dao.newContract
-          ? (sum.new += dao.newContractMembers.length)
-          : (sum.legacy += dao.apiDataStats.legacyData.members.length);
-        return sum;
-      },
-      { new: 0, legacy: 0 },
-    );
+  // const totalDoaMemberships = () => {
+  //   const counts = data.factories.reduce(
+  //     (sum, dao) => {
+  //       +dao.newContract
+  //         ? (sum.new += dao.newContractMembers.length)
+  //         : (sum.legacy += dao.apiDataStats.legacyData.members.length);
+  //       return sum;
+  //     },
+  //     { new: 0, legacy: 0 },
+  //   );
 
-    return counts.new + counts.legacy;
-  };
+  //   return counts.new + counts.legacy;
+  // };
 
-  const uniqueDaoMembers = () => {
-    const memberIDs = _.flatMap(data.factories, dao => {
-      return +dao.newContract
-        ? dao.newContractMembers
-        : dao.apiDataStats.legacyData.members;
-    }).map(member => {
-      return member.memberId ? member.memberId : member.id;
-    });
+  // const uniqueDaoMembers = () => {
+  //   const memberIDs = _.flatMap(data.factories, dao => {
+  //     return +dao.newContract
+  //       ? dao.newContractMembers
+  //       : dao.apiDataStats.legacyData.members;
+  //   }).map(member => {
+  //     return member.memberId ? member.memberId : member.id;
+  //   });
 
-    return _.uniq(memberIDs).length;
-  };
+  //   return _.uniq(memberIDs).length;
+  // };
 
   return (
     <div className="View Contain">
       <h1>BIG DAOta</h1>
-      {loading ? <p>Loading stats</p> : null}
+
+      <h2>
+        {' '}
+        <span role="img" aria-label="construction">
+          ⚒️
+        </span>
+        Under construction{' '}
+        <span role="img" aria-label="construction">
+          ⚒️
+        </span>
+      </h2>
+      {/* {loading ? <p>Loading stats</p> : null}
       {error ? <p>Error - are you on mainnet?</p> : null}
       {data ? (
         <div>
@@ -93,7 +104,7 @@ const Stats = props => {
             </TabPanel>
           </Tabs>
         </div>
-      ) : null}
+      ) : null} */}
     </div>
   );
 };
